+++ lib/api_request.dart
@@ -3,4 +3,5 @@ library api_request;
 export 'src/api_request.dart';
 export 'src/api_request_action.dart';
 export 'src/api_request_client.dart';
+export 'src/api_request_error.dart';
 export 'src/api_request_options.dart';