+++ lib/src/api_request_action.dart
@@ -1,14 +1,19 @@
+import 'dart:async';
+
+import 'package:api_request/src/interceptors/error_interceptor.dart';
 import 'package:dio/dio.dart';
 
 import 'api_request.dart';
 import 'api_request_client.dart';
+import 'api_request_error.dart';
 
 enum RequestMethod { GET, POST, PUT, DELETE }
 
 typedef ResponseBuilder<T> = T Function(dynamic);
 
-abstract class RequestAction<T, R extends ApiRequest?> {
+abstract class RequestAction<T, R extends ApiRequest> {
   final RequestClient? _requestClient = RequestClient.instance;
+  final StreamController<T> _streamController = StreamController<T>();
   R? request;
 
   bool get authRequired;
@@ -21,7 +26,30 @@ abstract class RequestAction<T, R extends ApiRequest?> {
 
   ResponseBuilder<T> get responseBuilder;
 
+  void onInit() {}
+
+  void onStart() {}
+
+  void onError(ApiRequestError error) {
+    if (!_streamController.isClosed) {
+      _streamController.sink.addError(error);
+      _streamController.close();
+    }
+  }
+
+  void onSuccess(T response) {
+    _streamController.sink.add(response);
+    _streamController.close();
+  }
+
+  StreamSubscription<T> onChange(
+      {Function(T response)? onSuccess, Function(Object error)? onError}) {
+    return _streamController.stream
+        .listen(onSuccess, cancelOnError: true, onError: onError);
+  }
+
   Future execute() async {
+    this.onStart();
     T _response;
     switch (this.method) {
       case RequestMethod.GET:
@@ -37,14 +65,34 @@ abstract class RequestAction<T, R extends ApiRequest?> {
         _response = responseBuilder(await delete());
         break;
     }
+    this.onSuccess(_response);
     return _response;
+  }
+
+  void onQueue() {
+    switch (this.method) {
+      case RequestMethod.GET:
+        get().then((value) => this.onSuccess(responseBuilder(value)));
+        break;
+      case RequestMethod.POST:
+        post().then((value) => this.onSuccess(responseBuilder(value)));
+        break;
+      case RequestMethod.PUT:
+        put().then((value) => this.onSuccess(responseBuilder(value)));
+        break;
+      case RequestMethod.DELETE:
+        delete().then((value) => this.onSuccess(responseBuilder(value)));
+        break;
+    }
   }
 
   var _dataMap;
 
-  RequestAction([this.request]) {
+  RequestAction(this.request) {
+    this.onInit();
     _requestClient?.configAuth(authRequired);
     handleRequest(this.request);
+    _requestClient?.addInterceptorOnce(ErrorInterceptor(this, this.onError));
   }
 
   Future<dynamic> get() async {